Monisha/repo/requirements.txt
absl-py==0.9.0
alembic==1.4.2
app==0.0.1
astor==0.8.1
astroid==2.4.0
autoenv==1.0.0
bcrypt==3.1.7
beautifulsoup4==4.8.0
bs4==0.0.1
cachetools==4.0.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cmake==3.15.3
cycler==0.10.0
Deprecated==1.2.10
dlib==19.19.0
et-xmlfile==1.0.1
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gast==0.2.2
get==2019.4.13
glob3==0.0.1
google-auth==1.10.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
html5lib==1.0.1
idna==2.8
imutils==0.5.3
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-vggface==0.6
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
Mako==1.1.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mtcnn==0.1.0
numpy==1.17.1
oauthlib==3.1.0
opencv-contrib-python==4.1.2.30
opencv-python==4.1.2.30
openpyxl==3.0.3
opt-einsum==3.1.0
pandas==0.25.1
pika==1.1.0
Pillow==7.0.0
post==2019.4.13
protobuf==3.11.2
psycopg2-binary==2.8.3
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyGithub==1.51
PyJWT==1.7.1
pylint==2.5.0
pyparsing==2.4.6
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
PyYAML==5.3
query-string==2019.4.13
rabbitmq==0.2.0
repoze.lru==0.7
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.3.0
Routes==2.4.1
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
selenium==3.141.0
six==1.12.0
sklearn==0.0
soupsieve==1.9.2
SQLAlchemy==1.3.8
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
toml==0.10.0
torch==1.3.1
typed-ast==1.4.0
urllib3==1.25.3
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
